requirements.txt
langgraph==0.2.0
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
arxiv==2.1.0
wikipedia==1.4.0
tavily-python==0.5.0
python-dotenv==1.0.1
streamlit==1.40.0
